+++ src/main/java/io/github/mzmine/modules/visualization/vankrevelendiagram/VanKrevelenDiagramXYZDataset.java
@@ -1,19 +1,19 @@
 /*
- * Copyright 2006-2020 The MZmine Development Team
- * 
+ * Copyright 2006-2021 The MZmine Development Team
+ *
  * This file is part of MZmine.
- * 
+ *
  * MZmine is free software; you can redistribute it and/or modify it under the terms of the GNU
  * General Public License as published by the Free Software Foundation; either version 2 of the
  * License, or (at your option) any later version.
- * 
+ *
  * MZmine is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even
- * the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General
- * Public License for more details.
- * 
+ * the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+ * General Public License for more details.
+ *
  * You should have received a copy of the GNU General Public License along with MZmine; if not,
- * write to the Free Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301
- * USA
+ * write to the Free Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
+ *
  */
 
 package io.github.mzmine.modules.visualization.vankrevelendiagram;
@@ -61,9 +61,9 @@ class VanKrevelenDiagramXYZDataset extends AbstractXYZDataset {
         if (zAxisLabel.equals("Retention time")) {
           zValuesList.add((double) filteredRows[i].getAverageRT());
         } else if (zAxisLabel.equals("Intensity")) {
-          zValuesList.add(filteredRows[i].getAverageHeight());
+          zValuesList.add(filteredRows[i].getAverageHeight().doubleValue());
         } else if (zAxisLabel.equals("Area")) {
-          zValuesList.add(filteredRows[i].getAverageArea());
+          zValuesList.add(filteredRows[i].getAverageArea().doubleValue());
         } else if (zAxisLabel.equals("Tailing factor")) {
           zValuesList.add((double) filteredRows[i].getBestFeature().getTailingFactor());
         } else if (zAxisLabel.equals("Asymmetry factor")) {
